chore: refactor campaign stats polling to a custom hook
Note: make `refreshCampaignStats` a memoized callback to satisfy the
useEffect dependencies.
